feat(issue-alert): add filter logic to previews (#40542)
Adds filter logic to previews. This pr adds `age_comparison`,
`issue_occurrences`, and `issue_category`. the other filter that we
currently have plans to support is `assigned_to`, but that requires some
more complicated assignee logic so it'll be added in a separate pr.

Filters have a `passes` method that is used to check if an incoming
`GroupEvent` will pass the filter. For previews we similarly have a
`passes_activity` method that checks if the filter passes a
`ConditionActivity`.