ref(savedsearch): Remove pinned search dedupe logic
This logic is likely from an err where the saved search frontend UI
caused this to be more problematic.

The result of this is simply that pinned searches that matched existing
saved searches now show up in the list next to your pinned search.

In the upcoming saved searches UI revamp we can simply reduplicate them
on the frontend. This really shouldn't be the responsibility of the
backend.