feat(replays): add count number to Replays tab inside Transaction Summary (#40798)
### Changes
- Added a fetch call to get the count number of replays of the
transaction in order to add the value to the Replays tab inside
Transaction Summary

Closes #40731 

### Test notes

- Made sure tests were passing
- Tested it with different transactions with/without replays

Loom videos: 


https://user-images.githubusercontent.com/39612839/199111548-c7c84bb7-3f49-4612-a79a-0625d95a8ae6.mp4



https://user-images.githubusercontent.com/39612839/199546303-bc80579f-4029-4332-8e90-4d4b264432e2.mp4



### Additional notes

- The header component is being unmounted and mounted every time the
user changes tabs, this result in the fetch call to be called again on
every tab change, making the number appear and disappear inside the
Replay tab.

- The count number will be slight off as this are counting replays that
are under 5 seconds which are being hidden in the table at the Replays
Tab. There are solutions for this but none of them are particularly
easy.