Limit creation of performance issues (#40621)
Limits creation of performance issues to grouphashes that have been seen
3 times in a 24 hour period. The third occurrence will cause the
creation of the issue and will count as the first seen event.

This will help cut down on noise from semi-parameterized query noise
until performance team is confident in their solution. This is counted
using an expiring redis key, so the time window has a theoretical
maximum of 48 hours.