fix: Fix post process forwarder metric (#40821)
Since this metric is a count, if the global SENTRY_METRICS setting is configured to anything other than 1 (which is the case at Sentry) this metric will be wrong.

This change now calls metrics.incr() with a sample rate of 1. Since we don't want to send too many metrics, we now buffer the counts and only flush them once per second.