feat(growth): sandbox walkthrough start guide (#40689)
this PR adds the start guide for the sandbox walkthrough. this is where
the guided tours begin and show you the different guided tours that are
available. changes to guideAnchors for guides have been made for the
sidebar/issue tours do not start immediately on arriving in the sandbox.
some copy has been changed so the same components can be used for
onboarding and the sandbox walkthrough.
<img width="399" alt="Screen Shot 2022-10-27 at 3 02 27 PM"
src="https://user-images.githubusercontent.com/46740234/198406757-f7fc5dff-f2e7-47cb-9cd4-d01fe69b404a.png">