fix(jira): Handle perf issue title (#40525)
Handle the title of a Jira ticket created from a performance issue. The
issue `title` isn't always available as it is for error issues and
renders as "untitled". This adds the type of issue to the title (called
the summary in Jira) e.g. `N+1 Query: SELECT "books_author"."id", ... `
to match what we render for error issues.

In the near future I want to add the span evidence as a table to the
Jira ticket description field, but that will be in a [follow up
PR](https://github.com/getsentry/sentry/pull/40535).