feat(spike-protection): Add hook for new settings field component (#40910)
See [ER-1231](https://getsentry.atlassian.net/browse/ER-1231)

This PR adds a hook for the new project settings field from getsentry.
It does a hack which I wasn't able to find a way around becuase of the
way the Form component overrides it's children's state without any
ability to opt out. That made it so that the toggle I was adding would
submit to the wrong endpoint. To get around this, I had to split the
form in two so that it wouldn't find the hooked component as a child,
and override it's submission params/endpoint.

This is also the best course of action since hooks should be independent
of where they are implemented in getsentry/sentry since they're in a
separate repository.