feat(growth): add functionality for demo walkthrough tasks (#40947)
this pr gives the ability to use onboarding tasks for the demo
walkthrough. these tasks are similar to the onboarding ones but are
unique to users, not just organizations. the middleware in the Sandbox
intercepts the api calls for the onboarding tasks and creates the
Sandbox specific onboarding tasks. However, because these onboarding
tasks are not stored in the Organization object, separate functions
(that are hooked in from the Sandbox) are needed to retrieve them so
they can be displayed in the sidebar and sidebar panel.

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>