feat(savedsearch): Use visibility in OrganizationSearchesEndpoint (take 2)
Revert for the revert c2f75d827d94852a417d24d87a034b71e659e9f5.

This is a follow up to https://github.com/getsentry/sentry/pull/40942

The code in this change is actually violating an existing constraint in
the production database that is **currently not correctly synced to the
django models** (whichg is why tests were passing).

Once we've corrected the constraints in production to properly match
what was supposed to be run in GH-40876 (but had to be skipped in
GH-40945) _then_ we can merge this.