feat(derived_code_mappings): Cache repository trees for orgs (#41016)
Github API requests have rate limiting. In order to prevent blowing through the rate limit we will be caching the trees for an org for 24 hours.

With the new post processing logic, we will be processing an event per hour per project. We can't recreate the repo trees per org on every request, otherwise, every event processed would use N + 1 requests (N is the number of repos for an org).