ref(serializer): Use select organization options as frontend feature flags (#41049)
This PR adopts a pattern I noticed for legacy ratelimiting of turning
select feature flags into flags passed through to the frontend. The
features list is great for selectively rendering UI, but doesn't work
with organization options out of the box. This makes there a single
source of truth so we don't have to manage a flagr flag alongside an org
option for rolling out orgoptions features.