fix(perf-issue): update Group.message value to use the title instead of transaction search_message (#41048)
When we create new Performance issues the `message` column gets
populated with the
[transaction.search_message](https://github.com/getsentry/sentry/blob/c2e29cd8493683c5479f48c7375e4e2443be5cd5/src/sentry/eventstore/models.py#L537-L565)
- which looks like an amalgamation of data at the event-level. For
error-issues, this works ok since the contextual 'culprit'/'problem'
would be available in the event data. But for transaction events, we run
the event through detection before we find a problem so the 'problem' is
decoupled from the transaction data.