fix(slack): Fix Slack level color for perf issues (#40416)
The lefthand sidebar of a Slack alert is colored based on the level of
the event, but performance issue alerts have a red sidebar (red=error)
and should have a blue sidebar (blue=info). Performance issues do have
the level set to info, so I suspect it's falling back to the default
value
[here](https://github.com/getsentry/sentry/blob/master/src/sentry/integrations/slack/message_builder/issues.py#L224)
as `event_for_tags` is None.