feat(jira): Add span evidence to Jira tickets for performance issues (#40535)
Add span evidence to the description of a Jira ticket created from a
performance issue. Currently the Jira issue is fairly empty as for an
error issue it shows the stacktrace, but for a performance issue it's
just a link back to the Sentry issue.

<img width="743" alt="Screen Shot 2022-11-03 at 3 40 10 PM"
src="https://user-images.githubusercontent.com/29959063/199847951-1b55a0f5-2313-4176-967b-0c09dff9f28c.png">

Note that I'm manually injecting a performance issue and the title is
displayed incorrectly here because it didn't go through that code path.
The point here is the description field.