feat(issue-alerts): incompatible AgeComparisonFilter and 'any' filterMatches (#41082)
Checks for incompatible `AgeComparisonFilter`s with
`FirstSeenEventCondition`. This is probably the last case that we're
missing.

This also adds checks for incompatible filters when the filter match is
`any` and all the filters are incompatible, and when the filter match is
`none`.

This also adds a lot more logic, but it could be cut down if we disallow
negative numbers in the form, since there's usually an alternative to
them.