ref(proj-config): Add context info to build task
Looking at transactions in the performance product, it's impossible to
know if tasks that take longer (including outliers) are all coming from
the same project or from different projects. Adding the tag should
address that.

This function is called in a single place and all the additional
parameters that go into `kwargs` is the time the task was scheduled,
which is not useful information, so I've decided to not include it (as
the context). We may want to revisit this in the future.