fix(pagerduty): dedupe on service not integration (#41137)
**context**
A user reported not getting pagerduty alerts sent to multiple services
for the same integration and sentry issue. At first this seemed to be
the cause of the `dedup_key` (PagerDuty has some
[docs](https://developer.pagerduty.com/docs/ZG9jOjExMDI5NTgx-sending-an-alert-event#alert-de-duplication)
on their alert deduplication.):

https://github.com/getsentry/sentry/blob/37ef54e8e9e5db280ccb7fe4eb6944b4bee609ed/src/sentry/integrations/pagerduty/client.py#L40


but after trying to cross date some `event_id`s with our pagerduty
"success" logs, it seemed like we weren't sending some of the
notifications it seemed we should have been sending.

**part of the fix**
I believe the problem in part may be due to the following: 

https://github.com/getsentry/sentry/blob/37ef54e8e9e5db280ccb7fe4eb6944b4bee609ed/src/sentry/rules/processor.py#L258-L261
we basically say that with a `future.key` we will group the rules
triggered together in one function (the callback) which makes sense if
you have different rules that are triggered for the same action, BUT
since the key for pagerduty is the `integration.id`, the future callback
is actually what's different and we should have two (one for each
service) in the case that you have two pagerduty actions in the same
rule.

**other part of the fix**
This PR will fix the above, but I believe the `dedup_key` may still come
into play. Maybe the `service.id` should be added to the `dedup_key` but
I think we should start with this and verify with the user who reported
this issue and the logs to see if we can verify it's truly on
pagerduty's end this time around

cc @Kobby-Bawuah