ref(release-health): Remove duplex backend from code (#41062)
In https://github.com/getsentry/getsentry/pull/8804, we switched SaaS
from using the duplex backend (which dispatches queries to either the
sessions or the metrics backend) to using the metrics backend
exclusively.

We are done comparing sessions and metrics results. The other use case
of the duplex backend was to have a rolling transition, allowing to use
the metrics dataset before we had 90 days of data. We assume we will not
need this transition in single-tenant. If we do, we can always write a
more simple dispatcher than the duplex backend.