test(relay): Instrument SDK initialization to catch race (#41176)
We observe infrequent test failures in the relay integration test suite. The
trigger is failed SDK initialization. After initializing the SDK, there is no
client bound to the hub. This should not happen, but we do not know whether the
client is reset or not bound in the first place.

To debug this, we plan to put assertions and a monitor in place that should give
us more information on where this problem originates.