fix(replays): decode replay as bytes if str from cache (#41231)
Trying to fix SENTRY-WDT, we first did #41111, which ensures that we
always receive the payload as bytes. however, depending on cache
configuration the cache can also return a string, which causes this
error to still happen. Given that, when we retrieve our segment part
from cache, we'll encode it if it's a string.

We'll likely want to compress this value for the cache which will always
render it un-encodable, so we can remove this check if we do that.