perf(events-stats): Use fromisoformat for events-stats zerofill
Following up to #41241. Turns out the date time parsing was indeed very slow in
the worst case.

Snuba returns a ISO 8601 timestamp using `datetime.isoformat()`. We parse it
using `dateutil.parser.parse`. This is extremely slow in the worst case
(10,000 timestamps), needing nearly a whole second. This change uses
`datetime.fromisoformat` introduced in Pytnon 3.7 to handle the parsing. This is
not a full ISO 8601 parser, rather it is the inverse function of
`datetime.isoformat()`. This is about 50-100X faster in the worst case.