perf(users-endpoint): Improve users endpoint performance
This endpoint relied on the query set to prefetch nested models from
OrganizationMember -> Team -> ProjectTeam -> Project. There's a lot of
redundancies in this many to many relationship so when prefetching, django is
doing a lot of processing to actually join the data in python. If there are
`a` members, each member is on `b` teams, and each team has `c` projects, then
django needs iterate `a` members, for each member, `b` teams, and for teach team
`c` projects, resulting in `a * b * c` operations in the worst case where every
member is on every team and every team has every project. This change aims to
improve this by splitting the query into 2 stages and manually perform the
lookups within the serializer instead of relying on django to construct all the
relationships and traversing them all.