ref(alert-preview): refactor preview code (#41278)
Refactors the preview code and separates all of the logic into separate
functions.

Other than that, the only difference is that instead of storing the
condition activity all in one list,
```
[<ConditionActivity1, group_id=1>, <ConditionActivity2, group_id=2>, <ConditionActivity3, group_id=1>]
```
We separate the activities by group_id
```
{
  1: [<ConditionActivity1>, <ConditionActivity3>],
  2: [<ConditionActivity2>],
}
```
This is done so that we can process the activity of each group
separately. This way we only need to store the data relevant to one
group at a time in memory to support frequency conditions.