fix: Add healthcheck for dev Kafka cluster
Sometime we have an issue when we proceed to running integration tests
before the Kafka cluster is actually fully started up.

This change adds additional healthcheck for kafka, but trying to list
the topics in the cluster and once we succeed we know that Kafka is
available now and we can start testing our code.