fix(tabs): Correctly identify overflowing tabs when adding/removing tabs (#41242)
This fixes a bug on the issues page where some tabs be put in the
overflow menu incorrectly.

The reason that this was happening was because the saved search tab was
toggling the `hidden` attribute. If that tab is `hidden={true}` on first
render, the ResizeObserver will not observe it since it's not being
rendered.

My solution is to always render all tabs and to add the `hidden`
attribute when they should not be shown. That way the ResizeObserver
will always be looking at the correct set of elements.