feat(perf-issues): Add boolean project options for issue creation
This adds a map of detector type to boolean project options which can be set via the UI so users have more granular control over detectors if they have noisy issues. Simultaneously I'm removing the project options for specific detector settings since it's complicates the detection code, and we already handle the tuning of this via system options. We can always re-introduce project level detector settings if it seems useful to offer to users later.