perf(users-endpoint): Improve users endpoint performance more (#41347)
The last attempt in #41292 made a small dent to the p75 but not
noticeable in the p95. Looking at it some more, the
`select_related("teams")` was fetching a many to many relationship and
reducing it in python which is currently the bottleneck. This change
removes all uses of `select_related` and `prefetch_related` entirely in
favour of a more granular control over the exact queries we make.

The total number of queries we make here is the same since we're just
breaking out the queries that would've been made by `prefetch_related`.
However, by manually constructing the final results from the queries,
we're able to avoid the work django does to add the attributes on the
models which was the true bottleneck here since we were working with 2
many to many relations here.