ref(options): Add allowlist to circumvent setup issues (#40401)
When we added the options permission blanketed across the entire
endpoint in the previous PR, it turns out to cause an issue in the
workflow where people 'make reset-db' and have a partially created user
(no super admin role yet) as they try to go through admin setup. This
allows only the admin email option (which was previously unpermissioned
for years anyway) to be updated in that flow.

It should help people setting up an account for the first time not run
into this error again. Either way, it's still an improvement over the
original as the endpoint is audit logged.