fix(profiling): Always use release render for profile details
This is a wild guess but there's a chance that organization isn't defined at
some point then we fall through to the default rendering of metadata values
which leads to the error.

The error itself seems to point to a completely different line but the message
led me to here.

Fixes JAVASCRIPT-2BA3