feat(digests): Enable sending digests with performance issues (#41463)
Add all issue types to email digests. Currently we hard code in a check
to see if it's an error issue and don't send them otherwise, but with
the performance issue launch we can add these in now. Here is what the
debug view looks like:

<img width="537" alt="Screen Shot 2022-11-15 at 5 00 22 PM"
src="https://user-images.githubusercontent.com/29959063/202285782-f2e4eb8b-fc9c-4ebc-8f80-8e3e6dc8c4b3.png">


The design may be updated, but this PR just enables the behavior. If we
want to wait for a redesign I can disable adding them in until that
time.