feat(profiling): remove implicit scoping to node (#41448)
This implicit scoping *sounded* smart, but in reality it just gets in
the way as we start to add more interactions. A click on a frame is
really not a strong enough indicator to warrant that pattern -
especially as folks often just click around the flamechart as they
explore it. This PR preserves all of the behavior we previously had
(like zooming into the frame) without scoping the call tree to that node
- the reducer state is kept as we think of the best way to expose that
action.