perf(projects): Correctly lazy load project cards [UP-212] (#41508)
Orgs with many projects have an unusable projects page due to the number
of rendered elements. There was a LazyLoad component, but it incorrectly
wrapped the entire results div, not individual project cards. This
rectifies that and makes the projects page usable again.

We should still probably paginate that page, but this is a quick fix
that at least makes things tenable in the meantime.