feat(indexer): Allow routing of messages (#40776)
### Context
In order to support slicing in Snuba, the upstream dependency of Snuba
i.e. Sentry metrics indexer needs to route messages to different slices.
The current implementation is hardcoded to send messages on a single
destination topic.

### Implementation
The way to support sending messages to different slices is by creating a
new abstraction called RoutingProducer. The routing producer handles all
the kafka specifics and delegates the routing decision to an abstract
`MessageRouter`. The `MessageRouter` returns a route which encapsulates
the producer and topic on which the data would be sent.

The `SlicingRouter` will use information about slicing from config and
populate all the sliced producers and topics. It will then rely on the
presence of `org_id` in the Kafkapayload header to make decisions on how
to route the message. The routing decision is similar to what is made in
Snuba.

**The current PR does not plug in the code with the consumer. There
would be a separate PR which would handle that.**

Co-authored-by: getsantry[bot] <66042841+getsantry[bot]@users.noreply.github.com>