ref(replays): Call useReplaysCount for each list of issues in the app (#41554)
This fixes an N+1 problem where the UI was making a request for each row in the Issue table. Previously it was requesting the count of replays that belong to the specific issue.
Now we're doing one request (per issues table) to get the count of replays for a list of issue ids.


Test Plan:
Issue table still works,
There's 1 request in the network tab, so all the counts pop in at the same time, see the highlighted area in the screen shot:

![Screen Shot 2022-11-17 at 8.27.26 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/ck6V58zD6DexpsEB0MNr/f6330e76-e4aa-4690-a37f-4c52e8e79cf6/Screen%20Shot%202022-11-17%20at%208.27.26%20PM.png)


Follows: https://github.com/getsentry/sentry/pull/41298

Fixes #40730