fix(saved-searches): Ensure that users can create saved searches with same query as other users (#41533)
Discovered an issue with the current logic:

1. User 1 in org 1 creates a personal saved search with query
`issue.category:performance`
2. User 2 in org 1 creates a personal saved search with query
`issue.category:performance`

This returns a 400 response currently, which is incorrect.

I simplified the checks to only disallow searches when they conflict
with one of the same type (i.e. an org search conflicting with an org
search or a user conflicting with the same user). In the past, org
searches could not conflict with a global search, but I do not see a
reason why we should disallow that (especially since we may modify
global searches in the future which could create conflicts anyway).