feat(issue-platform): Add `IssueOccurrence` class and add to `GroupEvent`
Adds in an `IssueOccurrence` class for representing details specific to a detected occurrence of an
issue. Also makes this an optional property of the `GroupEvent`.

Not completely clear on how this gets populated and passed around between processes yet. I'm
thinking that it probably makes sense to store this separately to `Event`, similar to what we did
for `PerformanceProblem`.