feat(onboarding): add default rule for creating project from onboarding (#41473)
If you create a project outside of onboarding we show an option to
create an issue alert for your project which is selected by default. But
our onboarding flow is more simple and doesn't have that. As a result,
we should just always create an alert rule when creating a new project.