feat(perf-issues) create base consecutive db detector (#41565)
The detector first looks for X number of consecutive db query spans,
each being at least Y in length, where X and Y are defined in the
detector threshold settings.
Once these set of spans are found, the detector will compare each db
span in the consecutive list
to determine if they are dependant on one another, if they are not a
performance issue is found.

The algorithm to determine if spans are independant. is quite simple to
begin, and fingerprinting is yet to be finalized. But the goal if this
PR is to run the detector in detection only mode.