fix(event_manager): Emit TRANSACTION outcomes if metrics are disabled (#41607)
In #40507 we started to count transaction metrics in the `transaction`
data category and transaction events in the `transaction_indexed` data
category. That PR missed that metrics extraction can be disabled, in
which case the old behavior of counting events as `transaction` should
remain. Relay already implemented this logic since getsentry/relay#1537
based on the metrics extraction flag.

This PR adds a feature check to the
`organizations:transaction-metrics-extraction` feature, which is the
same feature flag used to control Relay's behavior. We also remove the
previously used option to sample a percentage of organizations into
metrics extraction.

The default for this feature remains off (`false`) until metrics
components have been added to all deployment targets including
self-hosted.

Co-authored-by: Matej Minar <matej.minar@sentry.io>