feat(perf-issues): parameterize SAVEPOINTs in db span hashes
We found an instance in production of duplicate N+1s being created due to a
SAVEPOINT operation with a dynamic identifier. The example found was
specifically using double-quoted identifiers (PostgreSQL), but other possible
styles are unquoted (PostgreSQL and MySQL) or backtick-quoted (MySQL).

Replace SAVEPOINT identifiers with a placeholder for the purposes of hashing,
as two SAVEPOINT statements can be treated as logically equivalent for
performance issue detection/fingerprinting purposes.