feat(profiling): port fzf for search (#41744)
Fuse has been struggling with search and on long profiles we often start
blocking the main thread... Part of that seems to be because it tries to
find all occurrences, so likely runs in O(m*n) time where m is string
length and n is query length. We do not actually want it to match that
many cases as we've already seen that in some cases the matching was too
lax. Fzf seems to be a good substitute as it only ever finds the first
occurrence meaning it runs in O(n) time and is pretty fast. On top of
that, we also get fine control over search logic (for better or worse)
so we could always limit searching via raf with a time budget... For
now, this works well, I tested on the profiles that @eliashussary
pointed out and we do not seem to be dropping frames anymore