fix(hybrid-cloud): get_organizations does not query related data. (#41780)
I introduced get_organizations into the user serializer recently in my
continued attempts to enable more silo tests and convert code to hybrid
cloud compatibility by ensuring that organization objects (which exist
in the region silo) can be serialized into metadata from the user
serializer (in the control silo). The `get_organizations` service stub
represents a future silo RPC interaction to fulfill this need.

However, the existing `get_organizations` call was being used primarily
for `authentication` purposes, and required doing a large lookup for a
user's teams and other related data. Doing this across many users in the
user serializer therefore introduced large performance pain.

In reality, `get_organizations` should likely not be querying and
serializing related data. From the user serializer perspective, the need
is only on minor metadata such as the `slug` and `name`. From auth and
access, only the final id of "some" organization is needed, and once it
is determined that the teams data will be required, a separate, full
query can be made.

This should greatly improve the performance of both call sites.