feat(domains) Make slug updates work with customer domains (#41772)
When an org switches their slug & has subdomains enabled we need to
redirect them to their new domain. Unfortunately `browserHistory` is
unable to change history to a new domain so we need to use `.replace` on
window.location.

I've added a global mock for `window.location.replace` as we had it
defined in a few places and also have already mocked out several other
methods on `window.location`.