perf(search): yield to event loop when searching (#41758)
When searching through large profiles (10k+ frames) our for loop will
block the main thread and delay user input (easily 100ms+), this end up
making the whole page feeling very janky and unresponsive. Since we now
have granular control over both of our search loops we can wrap them in
a requestAnimationFrame and yield to the event loop when our ms budget
is crossed. This makes it less likely we will drop frames and not block
user input.

The budget is set to 12ms, so we have 4ms of whatever else may run to do
it's work.