feat(indexer): Fix indexer handling of backpressure
This attempts to fix how indexer is handling backpressure by keeping track
of whether it is in backpressure mode a boolean value `apply_backpressure`.
The previous implementation solved the problem by attempting to identify the
duplicate values in the batch by their offset number. But it wasn't partition
aware so if the batch spans multiple partitions it may have been wrong.

This is a required refactor for the upgrade to Arroyo 2.0.0, which will allow
messages to represent a batch of values from the broker than might span
multliple partitions.