bug(issues): Remove the cursor query param when navigating Issue tabs (#41719)
Fixes issue where switching in/out of the Issue>All Events tab can cause
query params to persist on the next page.

**Context**
These three tabs all use the `?cursor=` query param: 'All Events',
'Grouping', Replays'. When you switch between the tabs the cursor needs
to be reset; otherwise you could be looking at page 10 of the All Events
tab, but there aren't even 10 pages in the Grouping table.

**Current State**
To solve this problem the Grouping tab had two bits of code: the
`handleRouteLeave()` callback and also `updateUrlWithNewLevel()`. This
meant that moving _out_ of "Grouping" into another Tab would strip the
grouping specific query params. Also mvoing _into_ "Grouping" would
reset the params if both were not set, so Grouping could start over.

This worked when only All Events and Grouping used the same query param.
But now that Replays tab is here there are more combinations of to/from
links.
For example, a weird interaction that's happening right now is:
1. Visit 'All Events' Tab
  - Paginate to to the next set of data
2. Visit 'Merged Issues' Tab 
  - notice that `?cursor` is still in the url
3. Go back to 'All Events' 
  - notice that you're still on page 2!
4. Visit 'Grouping'
  - notice you're at page 1
- Paginate to the next set of data (might need use the "more issues"
slider)
5. Visit 'Merged Issues' again
  - notice that `?cursor` is _not_ in the url
6. Go back to 'Grouping'
  - notice that you're back on page 1


**What's changed**
This PR makes it so that each of these three pages adopts the
`handleRouteLeave()` method so that query params don't bleed from one
tab to the next.

All three tabs need to use this, but i'll add it to Replays in a
followup PR because there's some other issues with pagination there
right now. #merge-conflicts

Fixes #41791

Co-authored-by: Alberto Leal <mail4alberto@gmail.com>