fix(metrics): Use consume timestamp for metric outcomes (#41798)
Gets timestamps of outcomes derived from metrics closer to their
event-based counterpart. This should reduce a gap of usage numbers
between indexed and total transactions when transactions are backdated.

There are two types of performance traffic that are tracked with
separate outcomes:

- Transaction events flow through the `save_event` task. EventManager
  uses the `start_time` parameter to track outcomes for indexed
  transactions. That time is set by Relay during ingest and represents
  the time at which the event was first recorded at the ingest endpoint.
  This can be significantly different from the stated transaction
  timestamp if the transaction is submitted with a delay.
- Transaction metrics flow through the indexer and the
  `billing_metrics_consumer`. This consumer uses counts from the
  `transaction.duration` metric to record outcomes for total
  transactions. Metric buckets have no received timestamp, since the
  individual metrics that ended up in a bucket may have been received at
  different times. The former implementation of this consumer used the
  stated bucket timestamp to record outcomes, derived from the
  transaction timestamp. For backdated buckets, this means the outcome
  is recorded in the past rather than at ingest time.

This difference can create gaps where there are more indexed
transactions than total transactions in a given timeframe. Since metrics
do not have a `start_time` like events, it's not possible to avoid such
a gap entirely. However, the consumer can use the current time to
produce the metrics-based outcomes, which will be closer to the ingest
time than the bucket timestamp.