feat(profiling): allow jump to tree (#41686)
This PR add scroll to node functionality to useVirtualizedTree so that
we can scroll to nodes from different views. This enables us to scroll
to node in tree when user double clicks.


https://user-images.githubusercontent.com/9317857/203567862-e8dd94d0-3e27-433f-901f-f96dd7583b28.mp4

I've also taken the opportunity to also take some of the utils outside
of the useVirtualizedTree file as it was getting large, hence the big
diff here.