feat(replays): remove 5 second duration condition on replay snuba query (#41799)
We will test removing the `> 5 seconds` condition on our default query,
as this is filtering out replays that get tagged onto events. Even if
the replays in these cases are super short, they may contain valuable
information to help debug.

This will partly address our Replay Connectivity Issue
https://github.com/getsentry/replay-backend/issues/190, as right now
this is a source of "incorrect counts".